requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
cffi==1.14.1
cryptography==3.0
Django==3.1
django-bootstrap4==2.2.0
django-modeladmin-reorder==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.11.1
importlib-metadata==1.7.0
mysqlclient==2.0.1
Pillow==7.2.0
pkg-resources==0.0.0
pycparser==2.20
PyJWT==1.7.1
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
zipp==3.1.0
